+++ morpha/morpha/parsers/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from .irecord_parser import IRecordParser
+
+__all__ = ['IRecordParser']