+++ morpha/documentation/release_notes.md
@@ -2,4 +2,5 @@
 
 ## v0.1.0 (Pending)
 ##### Features
-- 
+- Added IRecordParser parser (`ME-630`).
+- Added TeamPreviewRecord model (`ME-630`).